+++ OpenProtocol/skyProtocolAdapter.m
@@ -42,7 +42,7 @@
 }
 
 // 适配器初始
-- (id)initAdapterWithProtocolType:(int)nProtocolType
+- (id)initAdapterWithProtocolType:(NSInteger)nProtocolType
 {
     self = [super init];
     if (self)
@@ -63,7 +63,7 @@
 }
 
 // 控制器连接
-- (BOOL)adapterConnectToController:(NSString *)ipAddress Port:(int)nPort
+- (BOOL)adapterConnectToController:(NSString *)ipAddress Port:(NSInteger)nPort
 {
     BOOL bResult = NO;
     // 协议判断
@@ -129,7 +129,7 @@
 
 #pragma mark - Protocol Adapter Interface
 // 1.控制器设置
-- (void)adapterInitialControllerRow:(int)nRow Column:(int)nColumn Resolution:(int)nRes
+- (void)adapterInitialControllerRow:(NSInteger)nRow Column:(NSInteger)nColumn Resolution:(NSInteger)nRes
 {
     // 协议判断
     switch (_adapterType) {
@@ -234,7 +234,7 @@
 }
 
 // 8.情景保存
-- (void)adapterSaveModelAtIndex:(int)nIndex
+- (void)adapterSaveModelAtIndex:(NSInteger)nIndex
 {
     // 协议判断
     switch (_adapterType) {
@@ -249,7 +249,7 @@
 }
 
 // 9.情景加载
-- (void)adapterLoadModelAtIndex:(int)nIndex
+- (void)adapterLoadModelAtIndex:(NSInteger)nIndex
 {
     // 协议判断
     switch (_adapterType) {
@@ -264,7 +264,7 @@
 }
 
 // 10.情景删除
-- (void)adapterDeleteModelAtIndex:(int)nIndex
+- (void)adapterDeleteModelAtIndex:(NSInteger)nIndex
 {
     // 协议判断
     switch (_adapterType) {
@@ -279,7 +279,7 @@
 }
 
 // 11.普通窗口信号切换
-- (void)adapterSignalSwitchSCXWin:(int)nWinID ofType:(int)nSrcType toChannel:(int)nSrcPath X:(int)nStartX Y:(int)nStartY H:(int)nHCount V:(int)nVCount;
+- (void)adapterSignalSwitchSCXWin:(NSInteger)nWinID ofType:(NSInteger)nSrcType toChannel:(NSInteger)nSrcPath X:(NSInteger)nStartX Y:(NSInteger)nStartY H:(NSInteger)nHCount V:(NSInteger)nVCount;
 {
     // 协议判断
     switch (_adapterType) {
@@ -294,7 +294,7 @@
 }
 
 // 12.叠加底图窗口信号切换
-- (void)adapterSignalSwitchOpenUnderWin:(int)nWinID ofType:(int)nSrcType toChannel:(int)nSrcPath
+- (void)adapterSignalSwitchOpenUnderWin:(NSInteger)nWinID ofType:(NSInteger)nSrcType toChannel:(NSInteger)nSrcPath
 {
     // 协议判断
     switch (_adapterType) {
@@ -309,7 +309,7 @@
 }
 
 // 13.叠加子窗口信号切换
-- (void)adapterSignalSwitchSubWin:(int)nWinID ofType:(int)nSrcType toChannel:(int)nSrcPath
+- (void)adapterSignalSwitchSubWin:(NSInteger)nWinID ofType:(NSInteger)nSrcType toChannel:(NSInteger)nSrcPath
 {
     // 协议判断
     switch (_adapterType) {
@@ -324,7 +324,7 @@
 }
 
 // 14.大画面拼接
-- (void)adapterSpliceSCXWin:(int)nWinID StartPanel:(int)nStart X:(int)nStartX Y:(int)nStartY V:(int)nVCount H:(int)nHCount ofType:(int)nSrcType toChannel:(int)nSrcPath
+- (void)adapterSpliceSCXWin:(NSInteger)nWinID StartPanel:(NSInteger)nStart X:(NSInteger)nStartX Y:(NSInteger)nStartY V:(NSInteger)nVCount H:(NSInteger)nHCount ofType:(NSInteger)nSrcType toChannel:(NSInteger)nSrcPath
 {
     // 协议判断
     switch (_adapterType) {
@@ -340,7 +340,7 @@
 }
 
 // 15.大画面分解
-- (void)adapterResolveSCXWin:(int)nWinID X:(int)nStartX Y:(int)nStartY H:(int)nHCount V:(int)nVCount
+- (void)adapterResolveSCXWin:(NSInteger)nWinID X:(NSInteger)nStartX Y:(NSInteger)nStartY H:(NSInteger)nHCount V:(NSInteger)nVCount
 {
     // 协议判断
     switch (_adapterType) {
@@ -355,7 +355,7 @@
 }
 
 // 16.进入叠加开窗
-- (void)adapterEnterOpenStatus:(int)nWinID ofType:(int)nSrcType toChannel:(int)nSrcPath
+- (void)adapterEnterOpenStatus:(NSInteger)nWinID ofType:(NSInteger)nSrcType toChannel:(NSInteger)nSrcPath
 {
     // 协议判断
     switch (_adapterType) {
@@ -370,7 +370,7 @@
 }
 
 // 17.退出叠加开窗
-- (void)adapterLeaveOpenStatus:(int)nWinID ofType:(int)nSrcType toChannel:(int)nSrcPath
+- (void)adapterLeaveOpenStatus:(NSInteger)nWinID ofType:(NSInteger)nSrcType toChannel:(NSInteger)nSrcPath
 {
     // 协议判断
     switch (_adapterType) {
@@ -385,7 +385,7 @@
 }
 
 // 18.添加子窗口
-- (void)adapterAddSubWin:(int)nSubID ofType:(int)nSrcType toChannel:(int)nSrcPath
+- (void)adapterAddSubWin:(NSInteger)nSubID ofType:(NSInteger)nSrcType toChannel:(NSInteger)nSrcPath
 {
     // 协议判断
     switch (_adapterType) {
@@ -400,7 +400,7 @@
 }
 
 // 19.关闭子窗口
-- (void)adapterDeleteSubWin:(int)nSubID ofType:(int)nSrcType toChannel:(int)nSrcPath
+- (void)adapterDeleteSubWin:(NSInteger)nSubID ofType:(NSInteger)nSrcType toChannel:(NSInteger)nSrcPath
 {
     // 协议判断
     switch (_adapterType) {
@@ -415,7 +415,7 @@
 }
 
 // 20.移动子窗口 同时包括位置与大小
-- (void)adapterMoveSubWin:(int)nSubID StartX:(int)nStartX StartY:(int)nStartY Width:(int)nWidth Height:(int)nHeight
+- (void)adapterMoveSubWin:(NSInteger)nSubID StartX:(NSInteger)nStartX StartY:(NSInteger)nStartY Width:(NSInteger)nWidth Height:(NSInteger)nHeight
 {
     // 协议判断
     switch (_adapterType) {
